Ramdisk and AnyKernel2 changes - Update AnyKernel2 to the latest available - Remove the whole fstab file and let AK2 do the work more cleanly - Update tools binaries - Add some dummy keys for possible boot verification (will test this later and see if it does anything at all)

+++ kernvatore.anykernel2/tools/ak2-core.sh
@@ -27,14 +27,61 @@ contains() { test "${1#*$2}" != "$1" && return 0 || return 1; }
 
 # dump boot and extract ramdisk
 dump_boot() {
-  dd if=$block of=/tmp/anykernel/boot.img;
-  $bin/unpackbootimg -i /tmp/anykernel/boot.img -o $split_img;
-  if [ $? != 0 ]; then
+  if [ ! -e "$(echo $block | cut -d\  -f1)" ]; then
+    ui_print " "; ui_print "Invalid partition. Aborting..."; exit 1;
+  fi;
+  if [ -f "$bin/nanddump" ]; then
+    $bin/nanddump -f /tmp/anykernel/boot.img $block;
+  else
+    dd if=$block of=/tmp/anykernel/boot.img;
+  fi;
+  if [ "$(strings /tmp/anykernel/boot.img | grep -E 'Green Loader|Green Recovery')" ]; then
+    mv -f /tmp/anykernel/boot.img /tmp/anykernel/boot-orig.img;
+    dd bs=1048576 count=1 conv=notrunc if=/tmp/anykernel/boot-orig.img of=$split_img/boot.img-master_boot.key;
+    dd bs=1048576 skip=1 conv=notrunc if=/tmp/anykernel/boot-orig.img of=/tmp/anykernel/boot.img;
+  fi;
+  if [ -f "$bin/unpackelf" -a "$($bin/unpackelf -i /tmp/anykernel/boot.img -h -q 2>/dev/null; echo $?)" == 0 ]; then
+    $bin/unpackelf -i /tmp/anykernel/boot.img -o $split_img;
+    mv -f $split_img/boot.img-ramdisk.cpio.gz $split_img/boot.img-ramdisk.gz;
+  elif [ -f "$bin/dumpimage" ]; then
+    $bin/dumpimage -l /tmp/anykernel/boot.img;
+    $bin/dumpimage -l /tmp/anykernel/boot.img > $split_img/boot.img-header;
+    grep "Name:" $split_img/boot.img-header | cut -c15- > $split_img/boot.img-name;
+    grep "Type:" $split_img/boot.img-header | cut -c15- | cut -d\  -f1 > $split_img/boot.img-arch;
+    grep "Type:" $split_img/boot.img-header | cut -c15- | cut -d\  -f2 > $split_img/boot.img-os;
+    grep "Type:" $split_img/boot.img-header | cut -c15- | cut -d\  -f3 | cut -d- -f1 > $split_img/boot.img-type;
+    grep "Type:" $split_img/boot.img-header | cut -d\( -f2 | cut -d\) -f1 | cut -d\  -f1 | cut -d- -f1 > $split_img/boot.img-comp;
+    grep "Address:" $split_img/boot.img-header | cut -c15- > $split_img/boot.img-addr;
+    grep "Point:" $split_img/boot.img-header | cut -c15- > $split_img/boot.img-ep;
+    $bin/dumpimage -i /tmp/anykernel/boot.img -p 0 $split_img/boot.img-zImage;
+    test $? != 0 && dumpfail=1;
+    if [ "$(cat $split_img/boot.img-type)" == "Multi" ]; then
+      $bin/dumpimage -i /tmp/anykernel/boot.img -p 1 $split_img/boot.img-ramdisk.gz;
+    else
+      dumpfail=1;
+    fi;
+  elif [ -f "$bin/rkcrc" ]; then
+    dd bs=4096 skip=8 iflag=skip_bytes conv=notrunc if=/tmp/anykernel/boot.img of=$split_img/boot.img-ramdisk.gz;
+  elif [ -f "$bin/pxa-unpackbootimg" ]; then
+    $bin/pxa-unpackbootimg -i /tmp/anykernel/boot.img -o $split_img;
+  else
+    $bin/unpackbootimg -i /tmp/anykernel/boot.img -o $split_img;
+  fi;
+  if [ $? != 0 -o "$dumpfail" ]; then
     ui_print " "; ui_print "Dumping/splitting image failed. Aborting..."; exit 1;
   fi;
   if [ -f "$bin/mkmtkhdr" ]; then
     dd bs=512 skip=1 conv=notrunc if=$split_img/boot.img-ramdisk.gz of=$split_img/temprd;
     mv -f $split_img/temprd $split_img/boot.img-ramdisk.gz;
+  fi;
+  if [ -f "$bin/unpackelf" -a -f "$split_img/boot.img-dtb" ]; then
+    case $(od -ta -An -N4 $split_img/boot.img-dtb | sed -e 's/del //' -e 's/   //g') in
+      QCDT|ELF) ;;
+      *) gzip $split_img/boot.img-zImage;
+         mv -f $split_img/boot.img-zImage.gz $split_img/boot.img-zImage;
+         cat $split_img/boot.img-dtb >> $split_img/boot.img-zImage;
+         rm -f $split_img/boot.img-dtb;;
+    esac;
   fi;
   mv -f $ramdisk /tmp/anykernel/rdtmp;
   mkdir -p $ramdisk;
@@ -49,37 +96,72 @@ dump_boot() {
 # repack ramdisk then build and write image
 write_boot() {
   cd $split_img;
-  cmdline=`cat *-cmdline`;
-  board=`cat *-board`;
-  base=`cat *-base`;
-  pagesize=`cat *-pagesize`;
-  kerneloff=`cat *-kerneloff`;
-  ramdiskoff=`cat *-ramdiskoff`;
-  tagsoff=`cat *-tagsoff`;
-  osver=`cat *-osversion`;
-  oslvl=`cat *-oslevel`;
-  if [ -f *-second ]; then
-    second=`ls *-second`;
-    second="--second $split_img/$second";
-    secondoff=`cat *-secondoff`;
-    secondoff="--second_offset $secondoff";
-  fi;
-  if [ -f /tmp/anykernel/zImage ]; then
-    kernel=/tmp/anykernel/zImage;
-  elif [ -f /tmp/anykernel/zImage-dtb ]; then
-    kernel=/tmp/anykernel/zImage-dtb;
+  if [ -f "$bin/mkimage" ]; then
+    name=`cat *-name`;
+    arch=`cat *-arch`;
+    os=`cat *-os`;
+    type=`cat *-type`;
+    comp=`cat *-comp`;
+    test "$comp" == "uncompressed" && comp=none;
+    addr=`cat *-addr`;
+    ep=`cat *-ep`;
   else
+    if [ -f *-cmdline ]; then
+      cmdline=`cat *-cmdline`;
+    fi;
+    if [ -f *-board ]; then
+      board=`cat *-board`;
+    fi;
+    base=`cat *-base`;
+    pagesize=`cat *-pagesize`;
+    kerneloff=`cat *-kerneloff`;
+    ramdiskoff=`cat *-ramdiskoff`;
+    if [ -f *-tagsoff ]; then
+      tagsoff=`cat *-tagsoff`;
+    fi;
+    if [ -f *-osversion ]; then
+      osver=`cat *-osversion`;
+    fi;
+    if [ -f *-oslevel ]; then
+      oslvl=`cat *-oslevel`;
+    fi;
+    if [ -f *-second ]; then
+      second=`ls *-second`;
+      second="--second $split_img/$second";
+      secondoff=`cat *-secondoff`;
+      secondoff="--second_offset $secondoff";
+    fi;
+    if [ -f *-hash ]; then
+      hash=`cat *-hash`;
+      test "$hash" == "unknown" && hash=sha1;
+      hash="--hash $hash";
+    fi;
+    if [ -f *-unknown ]; then
+      unknown=`cat *-unknown`;
+    fi;
+  fi;
+  for i in zImage zImage-dtb Image.gz Image Image-dtb Image.gz-dtb Image.bz2 Image.bz2-dtb Image.lzo Image.lzo-dtb Image.lzma Image.lzma-dtb Image.xz Image.xz-dtb Image.lz4 Image.lz4-dtb Image.fit; do
+    if [ -f /tmp/anykernel/$i ]; then
+      kernel=/tmp/anykernel/$i;
+      break;
+    fi;
+  done;
+  if [ ! "$kernel" ]; then
     kernel=`ls *-zImage`;
     kernel=$split_img/$kernel;
   fi;
-  if [ -f /tmp/anykernel/dtb ]; then
-    dtb="--dt /tmp/anykernel/dtb";
-  elif [ -f *-dtb ]; then
+  for i in dtb dt.img; do
+    if [ -f /tmp/anykernel/$i ]; then
+      dtb="--dt /tmp/anykernel/$i";
+      break;
+    fi;
+  done;
+  if [ ! "$dtb" -a -f *-dtb ]; then
     dtb=`ls *-dtb`;
     dtb="--dt $split_img/$dtb";
   fi;
   if [ -f "$bin/mkbootfs" ]; then
-    $bin/mkbootfs /tmp/anykernel/ramdisk | gzip > /tmp/anykernel/ramdisk-new.cpio.gz;
+    $bin/mkbootfs $ramdisk | gzip > /tmp/anykernel/ramdisk-new.cpio.gz;
   else
     cd $ramdisk;
     find . | cpio -H newc -o | gzip > /tmp/anykernel/ramdisk-new.cpio.gz;
@@ -87,19 +169,64 @@ write_boot() {
   if [ $? != 0 ]; then
     ui_print " "; ui_print "Repacking ramdisk failed. Aborting..."; exit 1;
   fi;
+  cd /tmp/anykernel;
   if [ -f "$bin/mkmtkhdr" ]; then
-    cd /tmp/anykernel;
     $bin/mkmtkhdr --rootfs ramdisk-new.cpio.gz;
     mv -f ramdisk-new.cpio.gz-mtk ramdisk-new.cpio.gz;
     case $kernel in
-      /tmp/anykernel/zImage*) $bin/mkmtkhdr --kernel $kernel; kernel=$kernel-mtk;;
+      $split_img/*) ;;
+      *) $bin/mkmtkhdr --kernel $kernel; kernel=$kernel-mtk;;
     esac;
   fi;
-  $bin/mkbootimg --kernel $kernel --ramdisk /tmp/anykernel/ramdisk-new.cpio.gz $second --cmdline "$cmdline" --board "$board" --base $base --pagesize $pagesize --kernel_offset $kerneloff --ramdisk_offset $ramdiskoff $secondoff --tags_offset $tagsoff --os_version "$osver" --os_patch_level "$oslvl" $dtb --output /tmp/anykernel/boot-new.img;
+  if [ -f "$bin/mkimage" ]; then
+    $bin/mkimage -A $arch -O $os -T $type -C $comp -a $addr -e $ep -n "$name" -d $kernel:$ramdisk boot-new.img;
+  elif [ -f "$bin/rkcrc" ]; then
+    $bin/rkcrc -k ramdisk-new.cpio.gz boot-new.img;
+  elif [ -f "$bin/pxa-mkbootimg" ]; then
+    $bin/pxa-mkbootimg --kernel $kernel --ramdisk ramdisk-new.cpio.gz $second --cmdline "$cmdline" --board "$board" --base $base --pagesize $pagesize --kernel_offset $kerneloff --ramdisk_offset $ramdiskoff $secondoff --tags_offset "$tagsoff" --unknown $unknown $dtb --output boot-new.img;
+  else
+    $bin/mkbootimg --kernel $kernel --ramdisk ramdisk-new.cpio.gz $second --cmdline "$cmdline" --board "$board" --base $base --pagesize $pagesize --kernel_offset $kerneloff --ramdisk_offset $ramdiskoff $secondoff --tags_offset "$tagsoff" --os_version "$osver" --os_patch_level "$oslvl" $hash $dtb --output boot-new.img;
+  fi;
   if [ $? != 0 ]; then
     ui_print " "; ui_print "Repacking image failed. Aborting..."; exit 1;
-  elif [ `wc -c < /tmp/anykernel/boot-new.img` -gt `wc -c < /tmp/anykernel/boot.img` ]; then
+  elif [ `wc -c < boot-new.img` -gt `wc -c < boot.img` ]; then
     ui_print " "; ui_print "New image larger than boot partition. Aborting..."; exit 1;
+  fi;
+  if [ -f "$bin/futility" -a -d "$bin/chromeos" ]; then
+    $bin/futility vbutil_kernel --pack boot-new-signed.img --keyblock $bin/chromeos/kernel.keyblock --signprivate $bin/chromeos/kernel_data_key.vbprivk --version 1 --vmlinuz boot-new.img --bootloader $bin/chromeos/empty --config $bin/chromeos/empty --arch arm --flags 0x1;
+    if [ $? != 0 ]; then
+      ui_print " "; ui_print "Signing image failed. Aborting..."; exit 1;
+    fi;
+    mv -f boot-new-signed.img boot-new.img;
+  fi;
+  if [ -f "$bin/BootSignature_Android.jar" -a -d "$bin/avb" ]; then
+    if [ -f "/system/system/bin/dalvikvm" ]; then
+      umount /system;
+      umount /system 2>/dev/null;
+      mkdir /system_root;
+      mount -o ro -t auto /dev/block/bootdevice/by-name/system$slot /system_root;
+      mount -o bind /system_root/system /system;
+    fi;
+    unset LD_LIBRARY_PATH;
+    pk8=`ls $bin/avb/*.pk8`;
+    cert=`ls $bin/avb/*.x509.*`;
+    /system/bin/dalvikvm -Xbootclasspath:/system/framework/core-oj.jar:/system/framework/core-libart.jar:/system/framework/conscrypt.jar:/system/framework/bouncycastle.jar -Xnodex2oat -Xnoimage-dex2oat -cp $bin/BootSignature_Android.jar com.android.verity.BootSignature /boot boot-new.img $pk8 $cert boot-new-signed.img;
+    if [ $? != 0 ]; then
+      ui_print " "; ui_print "Signing image failed. Aborting..."; exit 1;
+    fi;
+    mv -f boot-new-signed.img boot-new.img;
+    if [ -d "/system_root" ]; then
+      umount /system;
+      umount /system_root;
+      rmdir /system_root;
+      mount -o ro -t auto /system;
+    fi;
+  fi;
+  if [ -f "$bin/blobpack" ]; then
+    printf '-SIGNED-BY-SIGNBLOB-\00\00\00\00\00\00\00\00' > boot-new-signed.img;
+    $bin/blobpack tempblob LNX boot-new.img;
+    cat tempblob >> boot-new-signed.img;
+    mv -f boot-new-signed.img boot-new.img;
   fi;
   if [ -f "/data/custom_boot_image_patch.sh" ]; then
     ash /data/custom_boot_image_patch.sh /tmp/anykernel/boot-new.img;
@@ -107,11 +234,28 @@ write_boot() {
       ui_print " "; ui_print "User script execution failed. Aborting..."; exit 1;
     fi;
   fi;
-  dd if=/tmp/anykernel/boot-new.img of=$block;
+  if [ "$(strings /tmp/anykernel/boot.img | grep SEANDROIDENFORCE )" ]; then
+    printf 'SEANDROIDENFORCE' >> boot-new.img;
+  fi;
+  if [ -f "$bin/dhtbsign" ]; then
+    $bin/dhtbsign -i boot-new.img -o boot-new-signed.img;
+    mv -f boot-new-signed.img boot-new.img;
+  fi;
+  if [ -f "$split_img/boot.img-master_boot.key" ]; then
+    cat $split_img/boot.img-master_boot.key boot-new.img > boot-new-signed.img;
+    mv -f boot-new-signed.img boot-new.img;
+  fi;
+  if [ -f "$bin/flash_erase" -a -f "$bin/nandwrite" ]; then
+    $bin/flash_erase $block 0 0;
+    $bin/nandwrite -p $block /tmp/anykernel/boot-new.img;
+  else
+    dd if=/dev/zero of=$block 2>/dev/null;
+    dd if=/tmp/anykernel/boot-new.img of=$block;
+  fi;
 }
 
 # backup_file <file>
-backup_file() { cp $1 $1~; }
+backup_file() { test ! -f $1~ && cp $1 $1~; }
 
 # replace_string <file> <if search string> <original string> <replacement string>
 replace_string() {
@@ -125,7 +269,11 @@ replace_section() {
   begin=`grep -n "$2" $1 | head -n1 | cut -d: -f1`;
   for end in `grep -n "$3" $1 | cut -d: -f1`; do
     if [ "$begin" -lt "$end" ]; then
-      sed -i "/${2//\//\\/}/,/${3//\//\\/}/d" $1;
+      if [ "$3" == " " ]; then
+        sed -i "/${2//\//\\/}/,/^\s*$/d" $1;
+      else
+        sed -i "/${2//\//\\/}/,/${3//\//\\/}/d" $1;
+      fi;
       sed -i "${begin}s;^;${4}\n;" $1;
       break;
     fi;
@@ -137,7 +285,11 @@ remove_section() {
   begin=`grep -n "$2" $1 | head -n1 | cut -d: -f1`;
   for end in `grep -n "$3" $1 | cut -d: -f1`; do
     if [ "$begin" -lt "$end" ]; then
-      sed -i "/${2//\//\\/}/,/${3//\//\\/}/d" $1;
+      if [ "$3" == " " ]; then
+        sed -i "/${2//\//\\/}/,/^\s*$/d" $1;
+      else
+        sed -i "/${2//\//\\/}/,/${3//\//\\/}/d" $1;
+      fi;
       break;
     fi;
   done;
@@ -223,6 +375,28 @@ patch_fstab() {
     esac;
     newentry=$(echo "$entry" | sed "s;${part};${6};");
     sed -i "s;${entry};${newentry};" $1;
+  fi;
+}
+
+# patch_cmdline <cmdline match string> [<replacement string>]
+patch_cmdline() {
+  cmdfile=`ls $split_img/*-cmdline`;
+  if [ -z "$(grep "$1" $cmdfile)" ]; then
+    cmdtmp=`cat $cmdfile`;
+    echo "$cmdtmp $1" > $cmdfile;
+  else
+    match=$(grep -o "$1.*$" $cmdfile | cut -d\  -f1);
+    sed -i -e "s;${match};${2};" -e 's;  ; ;' -e 's;[ \t]*$;;' $cmdfile;
+  fi;
+}
+
+# patch_prop <prop file> <prop name> <new prop value>
+patch_prop() {
+  if [ -z "$(grep "^$2=" $1)" ]; then
+    echo -ne "\n$2=$3\n" >> $1;
+  else
+    line=`grep -n "^$2=" $1 | head -n1 | cut -d: -f1`;
+    sed -i "${line}s;.*;${2}=${3};" $1;
   fi;
 }
